+++ smmregrid/gridtype.py
@@ -4,16 +4,16 @@
 DEFAULT_DIMS = {
     'horizontal': ['i', 'j', 'x', 'y', 'lon', 'lat', 'longitude', 'latitude',
                    'cell', 'cells', 'ncells', 'values', 'value', 'nod2', 'pix', 'elem',
-                   'nav_lon', 'nav_lat'],
+                   'nav_lon', 'nav_lat', 'rgrid'],
     'vertical': ['lev', 'nz1', 'nz', 'depth', 'depth_full', 'depth_half'],
-    'time': ['time']
+    'time': ['time', 'time_counter'],
 }
 
 
 class GridType:
     """Fundamental GridType object"""
 
-    def __init__(self, dims, extra_dims=None, weights=None):
+    def __init__(self, dims, extra_dims=None, override=False, weights=None):
         """
         Initializes a GridType object carrying grid-specific information required by smmregrid.
 
@@ -21,6 +21,8 @@ class GridType:
             dims (list): A list of default dimensions for the grid (e.g., ['time', 'lat', 'lon']).
             extra_dims (dict, optional): A dictionary including keys 'vertical', 'time', and 'horizontal'
                                           that can be used to extend the default dimensions. Defaults to None.
+            override (bool, optional): If True, it will override the default dimensions with the provided ones.
+                                         Defaults to False.
             weights (any, optional): CDO weights used in regridding. It will initiate the object in a different
                                          way assuming single-gridtype objects. Defaults to None.
 
@@ -37,17 +39,15 @@ class GridType:
             weights_matrix (any): Placeholder for a weights matrix (to be defined later).
             level_index (str): A string used to identify the index of the levels.
 
-        Raises:
-            ValueError: If multiple vertical dimensions are identified during initialization.
         """
 
-        # key definitions
-        #if extra_dims:
-        #    for key, value in extra_dims.items():
-        #        if not isinstance(value, list):
-        #            raise ValueError(f'Extra_dims for {key} must be a list!')
+        # safety checks
+        #if not isinstance(dims, list):
+        #    raise TypeError("dims must be a list of dimension names.")
+        if extra_dims is not None and not isinstance(extra_dims, dict):
+            raise TypeError("extra_dims must be a dictionary or None.")
 
-        default_dims = self._handle_default_dimensions(extra_dims)
+        default_dims = self._handle_default_dimensions(extra_dims, override=override)
         self.horizontal_dims = self._identify_dims('horizontal', dims, default_dims)
         self.vertical_dim = self._identify_dims('vertical', dims, default_dims)
         self.dims = (self.horizontal_dims or []) + ([self.vertical_dim] if self.vertical_dim else [])
@@ -55,8 +55,9 @@ class GridType:
         self.other_dims = self._identify_other_dims(dims)
 
         # used by GridInspector
-        self.variables = {}
-        self.bounds = []
+        self.variables = {} # dictionary of variables and their coordinates
+        self.bounds = [] # list of bounds variables
+        self.kind = None  # which kind of grid, regular, guassian, curvilinear, etc.
 
         # used by Regrid class
         self.masked = None
@@ -64,7 +65,7 @@ class GridType:
         self.weights_matrix = None
         self.level_index = "idx_"
 
-    def _handle_default_dimensions(self, extra_dims):
+    def _handle_default_dimensions(self, extra_dims, override=False):
         """
         Extend the default dimensions based on the provided extra dimensions.
 
@@ -82,12 +83,24 @@ class GridType:
 
         if extra_dims is None:
             return DEFAULT_DIMS
+        
+        if override:
+            return extra_dims
 
         update_dims = DEFAULT_DIMS
         for dim in extra_dims.keys():
             if extra_dims[dim]:
                 update_dims[dim] = list(set(update_dims[dim] + extra_dims[dim]))
         return update_dims
+    
+    def __repr__(self):
+        """
+        Return a string representation of the GridType instance. 
+        Only includes attributes with values.
+        """
+
+        attributes = ', '.join(f"{key}={value}" for key, value in self.__dict__.items() if value is not None)
+        return f"{self.__class__.__name__}({attributes})"
 
     def __eq__(self, other):
         """
@@ -127,6 +140,16 @@ class GridType:
         Raises:
             ValueError: If more than one vertical dimension is identified.
         """
+
+        # Check if the axis is valid
+        if axis not in ['horizontal', 'vertical', 'time']:
+            raise ValueError(f"Invalid axis '{axis}'. Must be one of 'horizontal', 'vertical', or 'time'.")
+        
+        # Check if the axis is in the default dimensions
+        if axis not in default_dims:
+            return None
+        
+        # Identify dimensions based on the provided axis
         identified_dims = list(set(dims).intersection(default_dims[axis]))
         if axis == 'vertical':
             if len(identified_dims) > 1: